Move remove_fs tests
