+++ test/functional/android/appium_tests.py
@@ -13,11 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import base64
 import unittest
-from zipfile import ZipFile
 import os
-import random
 from time import sleep
 from dateutil.parser import parse
 
@@ -98,35 +95,6 @@ class AppiumTests(unittest.TestCase):
     def test_current_package(self):
         package = self.driver.current_package
         self.assertEqual('com.example.android.apis', package)
-
-    def test_push_pull_file(self):
-        path = '/data/local/tmp/test_push_file.txt'
-        data = b'This is the contents of the file to push to the device.'
-
-        self.driver.push_file(path, base64.b64encode(data).decode('utf-8'))
-        data_ret = base64.b64decode(self.driver.pull_file(path))
-
-        self.assertEqual(data, data_ret)
-
-    def test_pull_folder(self):
-        string_data = b'random string data %d' % random.randint(0, 1000)
-        path = '/data/local/tmp'
-
-        self.driver.push_file(path + '/1.txt', base64.b64encode(string_data).decode('utf-8'))
-        self.driver.push_file(path + '/2.txt', base64.b64encode(string_data).decode('utf-8'))
-
-        folder = self.driver.pull_folder(path)
-
-        # python doesn't have any functionality for unzipping streams
-        # save temporary file, which will be deleted in `tearDown`
-        self.zipfilename = 'folder_%d.zip' % random.randint(0, 1000000)
-        with open(self.zipfilename, "wb") as fw:
-            fw.write(base64.b64decode(folder))
-
-        with ZipFile(self.zipfilename, 'r') as myzip:
-            # should find these. otherwise it will raise a `KeyError`
-            myzip.read('1.txt')
-            myzip.read('2.txt')
 
     def test_background_app(self):
         self.driver.background_app(1)
