+++ test/functional/android/remote_fs_tests.py
@@ -14,13 +14,15 @@
 
 import base64
 import os
+import random
 import unittest
+from zipfile import ZipFile
 
 from appium import webdriver
 import desired_capabilities
 
 
-class PushFileTests(unittest.TestCase):
+class RemoteFsTests(unittest.TestCase):
     def setUp(self):
         desired_caps = desired_capabilities.get_desired_capabilities('ApiDemos-debug.apk')
         self.driver = webdriver.Remote('http://localhost:4723/wd/hub', desired_caps)
@@ -28,7 +30,36 @@ class PushFileTests(unittest.TestCase):
     def tearDown(self):
         self.driver.quit()
 
-    def test_push_file(self):
+    def test_push_pull_file(self):
+        path = '/data/local/tmp/test_push_file.txt'
+        data = b'This is the contents of the file to push to the device.'
+
+        self.driver.push_file(path, base64.b64encode(data).decode('utf-8'))
+        data_ret = base64.b64decode(self.driver.pull_file(path))
+
+        self.assertEqual(data, data_ret)
+
+    def test_pull_folder(self):
+        string_data = b'random string data %d' % random.randint(0, 1000)
+        path = '/data/local/tmp'
+
+        self.driver.push_file(path + '/1.txt', base64.b64encode(string_data).decode('utf-8'))
+        self.driver.push_file(path + '/2.txt', base64.b64encode(string_data).decode('utf-8'))
+
+        folder = self.driver.pull_folder(path)
+
+        # python doesn't have any functionality for unzipping streams
+        # save temporary file, which will be deleted in `tearDown`
+        self.zipfilename = 'folder_%d.zip' % random.randint(0, 1000000)
+        with open(self.zipfilename, "wb") as fw:
+            fw.write(base64.b64decode(folder))
+
+        with ZipFile(self.zipfilename, 'r') as myzip:
+            # should find these. otherwise it will raise a `KeyError`
+            myzip.read('1.txt')
+            myzip.read('2.txt')
+
+    def test_push_file_with_src_path(self):
         test_files = ['test_image.jpg', 'test_file.txt']
         for file_name in test_files:
             source_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), file_name)
@@ -43,5 +74,5 @@ class PushFileTests(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    suite = unittest.TestLoader().loadTestsFromTestCase(PushFileTests)
+    suite = unittest.TestLoader().loadTestsFromTestCase(RemoteFsTests)
     unittest.TextTestRunner(verbosity=2).run(suite)
